include

+++ scilab/modules/scicos/src/c/str2sci.c
@@ -21,6 +21,7 @@
 #include <string.h>
 #include "stack-c.h"
 #include "machine.h"
+#include "cvstr.h"
 
 void str2sci(char** x,int n,int m)
 {
@@ -51,7 +52,7 @@ void str2sci(char** x,int n,int m)
   *istk(il+3) = 0;
   *istk(il+4) = 1;
   for (i = 1; i <= n*m; i++){
-    n1=strlen(x[i-1]);
+    n1=(int)strlen(x[i-1]);
     *istk(il+4+i) =  *istk(il+4+i-1)+n1;
   if (n1 > 0) C2F(cvstr)(&n1,istk(il+m*n+5 -1 + *istk(il+3+i)),x[i-1],&zero,(unsigned long) n1);
   m1=m1+n1;
